Getting close  - todo in comments
Done with the core analyses and have some text in place for describing
SE math.  Still a bit fuzzy on it but it is there...

Also, got the sqr.year term to work OK, so it is incorporated.

Still TODO: Hand code robust clustered SE for 3b (there is a blog that
does this linked in the code...maybe adopt this?).

Also: The handwritten problems 1 and 2 are relatively weak right now.
I would say my versions get about 2/3 of the way there.  Matrix
operations are not my strong point.