Nightly commit.  Starting to understand...
It took me a while to wrap my mind around this assignment…I think I’m
there.  The big issue is that there doesn’t appear to be a relationship
between housing value (at any time) and the boundary for HRS.  Without
this the RD doesn’t work, right?  Am I missing something?  Maybe this
is the answer we’re supposed to get, it just seems very foreign to have
a “negative” result be correct in a class assignment, but hey maybe
that is a good lesson to give people…